code/orig_swc_process/requirements.txt
pyswcloader==1.1.1
pandas==1.5.3
numpy==1.23.5
pynrrd==1.1.3
tqdm==4.66.4
networkx==3.0
matplotlib==3.7.5